Move PTOLEDDisplay class out of __init__
